+++ src/lib.rs
@@ -1,6 +1,8 @@
+mod client;
 mod client_server;
 mod types;
 
+pub use client::WebClient;
 pub use client_server::{
     rocket, setup, CipherSubmission, DecryptionShareSubmission, DecryptionSharesMap,
     RegisteredUser, RegistrationOut, ServerResponse, TOTAL_USERS,
